backend/requirements.txt
# SciData Backend Dependencies

# Web 框架
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0

# 图像处理
Pillow>=10.0.0

# HTTP 客户端
httpx>=0.26.0

# 数据验证
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 数据库
supabase>=2.3.0

# AI
google-generativeai>=0.3.0

# 文件处理
python-multipart>=0.0.6
openpyxl>=3.1.0

# 环境变量
python-dotenv>=1.0.0

# 开发工具（可选）
# pytest>=7.0.0
# black>=23.0.0
# ruff>=0.1.0
